requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
bcrypt>=4.0.1
aiosqlite>=0.19.0
email-validator>=2.0.0
pydantic-settings>=2.0.0
asyncpg>=0.28.0
alembic>=1.12.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
httpx>=0.24.0 